feat: add very basic implementation
Adds a static implementation that takes a PID, sends SIGTERM to it,
waits for the process to exit, and sends SIGTERM after a user-specified
timeout if it doesn't.

Because the process identified by PID may not be a child process, the
wait syscall cannot be used. Instead, signal 0 is used to poll if the
process is running.

Signed-off-by: Ian Lewis <ian@ianlewis.org>